+++ xbmc/cores/RetroPlayer/rendering/VideoRenderers/RPRendererOpenGLES.cpp
@@ -7,6 +7,8 @@
  */
 
 #include "RPRendererOpenGLES.h"
+#include "cores/RetroPlayer/buffers/RenderBufferOpenGLES.h"
+#include "cores/RetroPlayer/buffers/RenderBufferPoolOpenGLES.h"
 #include "cores/RetroPlayer/rendering/RenderContext.h"
 #include "cores/RetroPlayer/rendering/RenderVideoSettings.h"
 #include "utils/GLUtils.h"
@@ -35,168 +37,6 @@ CRPBaseRenderer *CRendererFactoryOpenGLES::CreateRenderer(const CRenderSettings 
 RenderBufferPoolVector CRendererFactoryOpenGLES::CreateBufferPools(CRenderContext &context)
 {
   return { std::make_shared<CRenderBufferPoolOpenGLES>(context) };
-}
-
-// --- CRenderBufferOpenGLES ---------------------------------------------------
-
-CRenderBufferOpenGLES::CRenderBufferOpenGLES(CRenderContext &context,
-                                             GLuint pixeltype,
-                                             GLuint internalformat,
-                                             GLuint pixelformat,
-                                             GLuint bpp) :
-  m_context(context),
-  m_pixeltype(pixeltype),
-  m_internalformat(internalformat),
-  m_pixelformat(pixelformat),
-  m_bpp(bpp)
-{
-}
-
-CRenderBufferOpenGLES::~CRenderBufferOpenGLES()
-{
-  DeleteTexture();
-}
-
-void CRenderBufferOpenGLES::CreateTexture()
-{
-  glGenTextures(1, &m_textureId);
-
-  glBindTexture(m_textureTarget, m_textureId);
-
-  glTexImage2D(m_textureTarget, 0, m_internalformat, m_width, m_height, 0, m_pixelformat, m_pixeltype, NULL);
-
-  glTexParameteri(m_textureTarget, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-  glTexParameteri(m_textureTarget, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-  glTexParameteri(m_textureTarget, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
-  glTexParameteri(m_textureTarget, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
-
-  glBindTexture(m_textureTarget, 0);
-}
-
-bool CRenderBufferOpenGLES::UploadTexture()
-{
-  if (!glIsTexture(m_textureId))
-    CreateTexture();
-
-  glBindTexture(m_textureTarget, m_textureId);
-
-  const int stride = GetFrameSize() / m_height;
-
-  glPixelStorei(GL_UNPACK_ALIGNMENT, m_bpp);
-
-  if (m_bpp == 4 && m_pixelformat == GL_RGBA)
-  {
-    // XOR Swap RGBA -> BGRA
-    // GLES 2.0 doesn't support strided textures (unless GL_UNPACK_ROW_LENGTH_EXT is supported)
-    uint8_t* pixels = const_cast<uint8_t*>(m_data.data());
-    for (unsigned int y = 0; y < m_height; ++y, pixels += stride)
-    {
-      for (int x = 0; x < stride; x += 4)
-        std::swap(pixels[x], pixels[x + 2]);
-      glTexSubImage2D(m_textureTarget, 0, 0, y, m_width, 1, m_pixelformat, m_pixeltype, pixels);
-    }
-  }
-#ifdef GL_UNPACK_ROW_LENGTH_EXT
-  else if (m_context.IsExtSupported("GL_EXT_unpack_subimage"))
-  {
-    glPixelStorei(GL_UNPACK_ROW_LENGTH_EXT, stride / m_bpp);
-    glTexSubImage2D(m_textureTarget, 0, 0, 0, m_width, m_height, m_pixelformat, m_pixeltype, m_data.data());
-    glPixelStorei(GL_UNPACK_ROW_LENGTH_EXT, 0);
-  }
-#endif
-  else
-  {
-    uint8_t* pixels = const_cast<uint8_t*>(m_data.data());
-    for (unsigned int y = 0; y < m_height; ++y, pixels += stride)
-      glTexSubImage2D(m_textureTarget, 0, 0, y, m_width, 1, m_pixelformat, m_pixeltype, pixels);
-  }
-
-  glBindTexture(m_textureTarget, 0);
-
-  return true;
-}
-
-void CRenderBufferOpenGLES::DeleteTexture()
-{
-  if (glIsTexture(m_textureId))
-    glDeleteTextures(1, &m_textureId);
-
-  m_textureId = 0;
-}
-
-// --- CRenderBufferPoolOpenGLES -----------------------------------------------
-
-CRenderBufferPoolOpenGLES::CRenderBufferPoolOpenGLES(CRenderContext &context)
-  : m_context(context)
-{
-}
-
-bool CRenderBufferPoolOpenGLES::IsCompatible(const CRenderVideoSettings &renderSettings) const
-{
-  if (!CRPRendererOpenGLES::SupportsScalingMethod(renderSettings.GetScalingMethod()))
-    return false;
-
-  return true;
-}
-
-IRenderBuffer *CRenderBufferPoolOpenGLES::CreateRenderBuffer(void *header /* = nullptr */)
-{
-  return new CRenderBufferOpenGLES(m_context,
-                                   m_pixeltype,
-                                   m_internalformat,
-                                   m_pixelformat,
-                                   m_bpp);
-}
-
-bool CRenderBufferPoolOpenGLES::ConfigureInternal()
-{
-  switch (m_format)
-  {
-  case AV_PIX_FMT_0RGB32:
-  {
-    m_pixeltype = GL_UNSIGNED_BYTE;
-    if (m_context.IsExtSupported("GL_EXT_texture_format_BGRA8888") ||
-        m_context.IsExtSupported("GL_IMG_texture_format_BGRA8888"))
-    {
-      m_internalformat = GL_BGRA_EXT;
-      m_pixelformat = GL_BGRA_EXT;
-    }
-    else if (m_context.IsExtSupported("GL_APPLE_texture_format_BGRA8888"))
-    {
-      // Apple's implementation does not conform to spec. Instead, they require
-      // differing format/internalformat, more like GL.
-      m_internalformat = GL_RGBA;
-      m_pixelformat = GL_BGRA_EXT;
-    }
-    else
-    {
-      m_internalformat = GL_RGBA;
-      m_pixelformat = GL_RGBA;
-    }
-    m_bpp = sizeof(uint32_t);
-    return true;
-  }
-  case AV_PIX_FMT_RGB555:
-  {
-    m_pixeltype = GL_UNSIGNED_SHORT_5_5_5_1;
-    m_internalformat = GL_RGB;
-    m_pixelformat = GL_RGB;
-    m_bpp = sizeof(uint16_t);
-    return true;
-  }
-  case AV_PIX_FMT_RGB565:
-  {
-    m_pixeltype = GL_UNSIGNED_SHORT_5_6_5;
-    m_internalformat = GL_RGB;
-    m_pixelformat = GL_RGB;
-    m_bpp = sizeof(uint16_t);
-    return true;
-  }
-  default:
-    break;
-  }
-
-  return false;
 }
 
 // --- CRPRendererOpenGLES -----------------------------------------------------
