+++ xbmc/cores/RetroPlayer/rendering/VideoRenderers/RPRendererOpenGLES.h
@@ -35,62 +35,6 @@ namespace RETRO
     RenderBufferPoolVector CreateBufferPools(CRenderContext &context) override;
   };
 
-  class CRenderBufferOpenGLES : public CRenderBufferSysMem
-  {
-  public:
-    CRenderBufferOpenGLES(CRenderContext &context,
-                          GLuint pixeltype,
-                          GLuint internalformat,
-                          GLuint pixelformat,
-                          GLuint bpp);
-    ~CRenderBufferOpenGLES() override;
-
-    // implementation of IRenderBuffer via CRenderBufferSysMem
-    bool UploadTexture() override;
-
-    GLuint TextureID() const { return m_textureId; }
-
-  protected:
-    // Construction parameters
-    CRenderContext &m_context;
-    const GLuint m_pixeltype;
-    const GLuint m_internalformat;
-    const GLuint m_pixelformat;
-    const GLuint m_bpp;
-
-    const GLenum m_textureTarget = GL_TEXTURE_2D; //! @todo
-    GLuint m_textureId = 0;
-
-    void CreateTexture();
-
-  private:
-    void DeleteTexture();
-  };
-
-  class CRenderBufferPoolOpenGLES : public CBaseRenderBufferPool
-  {
-  public:
-    CRenderBufferPoolOpenGLES(CRenderContext &context);
-    ~CRenderBufferPoolOpenGLES() override = default;
-
-    // implementation of IRenderBufferPool via CBaseRenderBufferPool
-    bool IsCompatible(const CRenderVideoSettings &renderSettings) const override;
-
-  protected:
-    // implementation of CBaseRenderBufferPool
-    IRenderBuffer *CreateRenderBuffer(void *header = nullptr) override;
-    bool ConfigureInternal() override;
-
-    // Construction parameters
-    CRenderContext &m_context;
-
-    // Configuration parameters
-    GLuint m_pixeltype = 0;
-    GLuint m_internalformat = 0;
-    GLuint m_pixelformat = 0;
-    GLuint m_bpp = 0;
-  };
-
   class CRPRendererOpenGLES : public CRPBaseRenderer
   {
   public:
